Fixing non-keyed/stdweb to compile
