+++ frameworks/non-keyed/stdweb/Cargo.toml
@@ -3,7 +3,11 @@ name = "js-framework-benchmark-stdweb"
 version = "1.0.0"
 authors = ["Isamu Mogi <isamu@leafytree.jp>"]
 
+[profile.release]
+lto = true
+panic = "abort"
+
 [dependencies]
 byteorder = { version = "1.2.4", features = [ "i128" ] }
 rand = { version = "0.5.4", features = [ "stdweb" ] }
-stdweb = "0.4.8"
+stdweb = "0.4.17"
