+++ frameworks/non-keyed/stdweb/README.md
@@ -1,0 +1,12 @@
+Because of the heavy amount of Rust dependencies, this example is pre-compiled, so you don't need to compile anything.
+
+However, if you do want to compile it, you will need the following:
+
+* [Rust](https://www.rust-lang.org/tools/install)
+
+After installing that, run these commands:
+
+```
+npm install
+npm run build-prod-force
+```
