--- frameworks/non-keyed/stdweb/Web.toml
@@ -1,1 +1,0 @@
-default-target = "wasm32-unknown-unknown"
