+++ frameworks/non-keyed/stdweb/bundled-dist/index.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=1" name="viewport" />
+    <script>
+        var Module = {};
+        var __cargo_web = {};
+        Object.defineProperty( Module, 'canvas', {
+            get: function() {
+                if( __cargo_web.canvas ) {
+                    return __cargo_web.canvas;
+                }
+
+                var canvas = document.createElement( 'canvas' );
+                document.querySelector( 'body' ).appendChild( canvas );
+                __cargo_web.canvas = canvas;
+
+                return canvas;
+            }
+        });
+    </script>
+</head>
+<body>
+    <script src="js-framework-benchmark-stdweb.js"></script>
+</body>
+</html>
