+++ frameworks/non-keyed/stdweb/index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8"/>
     <title>stdweb</title>
     <link href="/css/currentStyle.css" rel="stylesheet"/>
+    <base href="bundled-dist/"></base>
 </head>
 <body>
 <div id="main">
