+++ frameworks/non-keyed/stdweb/package.json
@@ -4,16 +4,18 @@
   "description": "Benchmark for stdweb",
   "license": "ISC",
   "js-framework-benchmark": {
-    "frameworkVersion": "0.4.8"
+    "frameworkVersion": "0.4.17"
   },
   "scripts": {
     "build-prod": "echo This is a no-op. && echo Due to heavy dependencies, the generated javascript is already provided. && echo If you really want to rebuild from source use: && echo npm run build-prod-force",
-    "build-dev-force": "node build.js web deploy --verbose",
-    "build-prod-force": "node build.js web deploy --verbose --release",
-    "start": "node build.js web start --verbose --open --auto-reload"
+    "build-prod-force": "rimraf bundled-dist target/deploy && rustup toolchain install nightly && rustup override set nightly && rustup target add wasm32-unknown-unknown && cargo install -f cargo-web && cargo web deploy --release --target wasm32-unknown-unknown && cpr target/deploy bundled-dist"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/krausest/js-framework-benchmark.git"
+  },
+  "devDependencies": {
+    "rimraf": "^2.6.3",
+    "cpr": "^3.0.1"
   }
 }
