--- frameworks/non-keyed/stdweb/rust-toolchain
@@ -1,1 +1,0 @@
-nightly-2018-07-11
