--- frameworks/non-keyed/stdweb/static/index.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8"/>
-    <title>stdweb</title>
-    <link href="/css/currentStyle.css" rel="stylesheet"/>
-</head>
-<body>
-<div id="main">
-    <div class="container">
-        <div class="jumbotron">
-            <div class="row">
-                <div class="col-md-6">
-                    <h1>stdweb</h1>
-                </div>
-                <div class="col-md-6">
-                    <div class="row">
-                        <div class="col-sm-6 smallpad">
-                            <button type="button" class="btn btn-primary btn-block" id="_run">Create 1,000 rows</button>
-                        </div>
-                        <div class="col-sm-6 smallpad">
-                            <button type="button" class="btn btn-primary btn-block" id="_runlots">Create 10,000 rows</button>
-                        </div>
-                        <div class="col-sm-6 smallpad">
-                            <button type="button" class="btn btn-primary btn-block" id="_add">Append 1,000 rows</button>
-                        </div>
-                        <div class="col-sm-6 smallpad">
-                            <button type="button" class="btn btn-primary btn-block" id="_update">Update every 10th row</button>
-                        </div>
-                        <div class="col-sm-6 smallpad">
-                            <button type="button" class="btn btn-primary btn-block" id="_clear">Clear</button>
-                        </div>
-                        <div class="col-sm-6 smallpad">
-                            <button type="button" class="btn btn-primary btn-block" id="_swaprows">Swap Rows</button>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <table class="table table-hover table-striped test-data">
-            <tbody id="tbody">
-            </tbody>
-        </table>
-        <span class="preloadicon glyphicon glyphicon-remove" aria-hidden="true"></span>
-    </div>
-</div>
-<script src="js-framework-benchmark-stdweb.js"></script>
-</body>
-</html>
